// types.d.ts
export interface ExploreType {
    img: string
    location: string
    distance: string
}

export interface LivingType {
    img: string
    title: string
}

export interface StayType {
    img: string
    location: string
    title: string
    description: string
    star: number
    price: string
    total: string
    long: number
    lat: number
}